requirements.txt
streamlit
replicate
openai
pandas
folium
pickle
streamlit-authenticator
streamlit-extra
streamlit-card
streamlit-folium